telegram-solana-sales-bot/requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot
python-dotenv
jinja2
python-multipart
